+++ packages/ignition/src/app/components/icons/icons-collection/use-icon-change.ts
@@ -1,0 +1,17 @@
+import { useState, useEffect } from "react";
+import { useParams } from "next/navigation";
+
+const useIconChange = (): {
+  selected: string;
+} => {
+  const params = useParams<{ iconid?: string }>();
+  const [selected, setSelected] = useState<string>("");
+
+  useEffect(() => {
+    setSelected(params?.iconid ?? "");
+  }, [params?.iconid]);
+
+  return { selected };
+};
+
+export default useIconChange;